Cross-build most of the tools.
