+++ tests/build.py
@@ -1,5 +1,6 @@
 from build.ab2 import normalrule, Rule, Target, export
-from src.build import TOOLCHAINS, cowgol
+from src.build import cowgol
+from src.toolchains import TOOLCHAINS
 
 TESTS = [
     "addsub-16bit",
